chore(#1169): first cucumber test for landing and wizard
